A whole lotta stuff (read Description)
So I added a lovely little sayall function which will print out a message in the chat using the THERMAL tag so people know it's the server.

we have a on damage function which is called when a player takes damage.

we have a melee checker, if the damage was taken via melee then a message it displayed, if it was not melee then nothing is displayed and the player takes no damage (useful against preventing gun kills)