I done alot, read description of commit
Changed file name to something easier to type in console and better reflects the actual gamemode name.

Changed the messages printed when the game is loading and booting up.

And other updates that I'm too tired to remember ( It's 10:11AM and I stayed up till 2am last night so I'm tired af, not to mention i got a LOT of work todo today )